Modernize overrides. (#11)
Using more recent versions of Ruby means we can:

* Get rid of the `send :include` hack. `include` is now public.
* Get rid of `alias_method_chain`. `prepend` is more flexible.
* Remove the `class_eval` and instead prepend on the `singleton_class`

In addition I:

* Used `ActiveSupport::Concern` to reduce the boilerplate. Since we are
  already using ActionMailer, `ActiveSupport::Concern` will be avialable.
* Renamed some of the modules to be more relevant.

The primary reason for this commit was to replace the `alias_method_chain`
with `prepend`. I'm want the request info to be used in Mandrill templated
e-mails as well. I'm using `mandrill_mailer` gem which has a ActionMailer-like
API. The default url options override that this gem provides couldn't be applied
there because `default_url_options` is not a real method in mandrill_mailer
so `alias_method_chain` doesn't work. The more modern `prepend` does work. This
commit allows me to get use it for my purposes as well as modernize this gem a
bit.

For those wanting to also use it for the `mandrill_mailer` gem you just need to
add the following to your initializer if using this commit:

    Rails.application.config.mandrill_mailer.default_url_options = { host: 'yourdomain.com' }
    MandrillMailer.config.singleton_class.prepend ActionMailerWithRequest::DefaultUrlOptionsOverride